requirements.txt
Flask==3.1.0
Pillow==10.2.0 
tensorflow==2.19.0
numpy==1.26.3